requirements.txt
diffusers==0.33.1
transformers==4.51.3
pynvml==12.0.0
accelerate==1.12.0
safetensors==0.7.0
pytest==9.0.2
pytest-mock==3.15.1
gradio==5.34.1
numpy==2.2.5
pillow==11.2.1
futures==3.0.5
sentencepiece==0.2.0
sacremoses==0.1.1
einops==0.8.1
timm==1.0.15
scipy==1.15.2
torchsde==0.2.6
tqdm==4.67.1
peft==0.15.2
compel==2.0.3
modelscope==1.25.0
huggingface_hub==0.30.2
onnxruntime-gpu==1.21.1
rembg[gpu]==2.0.65
piexif==1.1.3
opencv-python-headless==4.11.0.86
datasets==2.16.1
simplejson==3.20.1
sortedcontainers==2.4.0
scikit-image==0.25.2
addict==2.4.0
bitsandbytes==0.46.0
vtracer==0.6.11
